third_party/blink/web_tests/platform/generic/fast/forms/selection-setSelectionRange-focusing-expected.txt

Test for crbug.com/393504 :HTMLTextAreaElement.setSelectionRange should not change focus.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.activeElement is document.body
PASS document.activeElement is document.body
PASS document.activeElement is document.body
PASS document.activeElement is document.body
PASS document.activeElement is document.body
PASS successfullyParsed is true

TEST COMPLETE

 
